backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "doctr>=1.9.0",
    "fastapi>=0.120.4",
    "fhir-resources>=8.1.0",
    "gemini-ai>=0.0.5",
    "google-generativeai>=0.8.5",
    "ipfshttpclient>=0.7.0",
    "pdfminer-six>=20250506",
    "pydantic>=2.12.3",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "scispacy>=0.6.2",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "unstructured>=0.18.15",
    "uuid>=1.30",
    "uvicorn[standard]>=0.38.0",
    "web3>=7.14.0",
]
